+++ client/packages/common/src/hooks/useNativeClient/useNativeClient.ts
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react';
 import { registerPlugin, Capacitor } from '@capacitor/core';
 import { useLocalStorage } from '../../localStorage';
 
-const DISCOVERY_TIMEOUT = 5000;
+const DISCOVERY_TIMEOUT = 7000;
 const DISCOVERED_SERVER_POLL = 2000;
 export const PREVIOUS_SERVER_KEY = '/discovery/previous-server';
 export const NATIVE_MODE_KEY = '/native/mode';
