+++ server/android/Cargo.toml
@@ -22,3 +22,5 @@ server = { path = "../server", default-features = false, features = [
     "android",
 ] }
 tokio = { version = "1.17.0", features = ["macros"] }
+android_logger = { version = "0.13.0" }
+fast_log = { version = "1.5" }
