+++ server/server/Cargo.toml
@@ -50,7 +50,6 @@ mime_guess = "2.0.4"
 futures = "0.3"
 fast_log = { version = "1.5" }
 rcgen = "0.9.2"
-android_logger = { version = "0.13.0", optional = true }
 
 
 [dev-dependencies]
@@ -69,4 +68,4 @@ default = ["sqlite", "dep:machine-uid"]
 sqlite = ["repository/sqlite"]
 postgres = ["repository/postgres"]
 # See openssl comment regarding the "openssl/vendored" feature
-android = ["repository/sqlite", "openssl/vendored", "dep:android_logger"]
+android = ["repository/sqlite", "openssl/vendored"]
