+++ server/server/src/logging.rs
@@ -1,50 +1,33 @@
 use std::env;
 
 use fast_log::{
-    appender::{FastLogRecord, LogAppender},
     consts::LogSize,
     plugin::{file_split::RollingType, packer::LogPacker},
     Config as LogConfig,
 };
-use log::{LevelFilter, Record};
+use log::LevelFilter;
 use service::settings::{LogMode, LoggingSettings};
 
-pub fn logging_init(settings: Option<LoggingSettings>) {
+pub fn logging_init(
+    settings: Option<LoggingSettings>,
+    apply_config: Option<Box<dyn Fn(LogConfig) -> LogConfig>>,
+) {
     let settings = settings.unwrap_or(LoggingSettings::new(
         LogMode::Console,
         service::settings::Level::Info,
     ));
-    let config = match settings.mode {
+    let mut config = match settings.mode {
         LogMode::File => file_logger(&settings),
         LogMode::Console => LogConfig::new().console(),
         LogMode::All => file_logger(&settings).console(),
     };
-    fast_log::init(
-        config
-            .level(LevelFilter::from(settings.level.clone()))
-            .custom(NativeLogger {}),
-    )
-    .expect("Unable to initialise logger");
-    log::warn!("omSupply Logging initialised aaabbbccc");
-}
 
-struct NativeLogger {}
-impl LogAppender for NativeLogger {
-    fn do_logs(&self, records: &[FastLogRecord]) {
-        #[cfg(feature = "android")]
-        {
-            // logs to the android logcat in addition to the standard oms log file
-            records.iter().for_each(|record| {
-                android_logger::log(
-                    &Record::builder()
-                        .args(format_args!("{}", record.args))
-                        .target("omSupply")
-                        .level(record.level)
-                        .build(),
-                )
-            });
-        }
+    if let Some(apply_config) = apply_config {
+        config = apply_config(config);
     }
+
+    fast_log::init(config.level(LevelFilter::from(settings.level.clone())))
+        .expect("Unable to initialise logger");
 }
 
 fn file_logger(settings: &LoggingSettings) -> LogConfig {
