+++ server/server/src/main.rs
@@ -8,7 +8,7 @@ use service::settings::Settings;
 async fn main() -> std::io::Result<()> {
     let settings: Settings =
         configuration::get_configuration().expect("Failed to parse configuration settings");
-    logging_init(settings.logging.clone());
+    logging_init(settings.logging.clone(), None);
 
     let off_switch = tokio::sync::mpsc::channel(1).1;
     start_server(settings, off_switch).await
